+++ src/main/java/ba/atlantbh/auctionapp/repositories/BidRepository.java
@@ -6,6 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.UUID;
 
@@ -16,5 +17,5 @@ public interface BidRepository extends JpaRepository<Bid, UUID> {
     List<SimpleBidResponse> getBidsForProduct(String id);
 
     @Query(value = "SELECT MAX(price) FROM bid WHERE person_id = :person_id AND product_id = :product_id", nativeQuery = true)
-    Float getMaxBidFromPersonForProduct(@Param("person_id") String personId, @Param("product_id") String productId);
+    BigDecimal getMaxBidFromPersonForProduct(@Param("person_id") String personId, @Param("product_id") String productId);
 }
